Finished Week 5
Finished week 5

guessing game: allows user to enter a guess and then tells them if they
are high or low. Background color changes to reflect this as well.

TextFileIO: Creates a numbers.dat file, saves even numbers 2-100 to the
file, closes the file and displays it. Then reopens file, appends odd
numbers to the file and redisplays the file.